requirements.txt
fastapi
uvicorn
python-dotenv
requests
python-multipart
pydantic
PyPDF2==3.0.1
python-docx==1.1.0
